Changed disc_read_file method's return type to size_t
Failed read is indicated with zero return value

+++ src/libbluray/disc/disc.h
@@ -57,7 +57,7 @@ BD_PRIVATE struct bd_file_s *disc_open_path(BD_DISC *disc, const char *path);
 BD_PRIVATE struct bd_dir_s *disc_open_dir (BD_DISC *disc, const char *dir);
 
 /* Read VFS file */
-BD_PRIVATE int64_t disc_read_file(BD_DISC *disc, const char *dir, const char *file,
+BD_PRIVATE size_t disc_read_file(BD_DISC *disc, const char *dir, const char *file,
                                   uint8_t **data);
 
 /* Update virtual package */
